Refactor "colour.io.luts.lut" module.
